+++ stroom-core-shared/src/main/java/stroom/data/shared/StreamTypeNames.java
@@ -1,5 +1,8 @@
 package stroom.data.shared;
 
+import java.util.HashSet;
+import java.util.Set;
+
 public class StreamTypeNames {
 
     /**
@@ -50,4 +53,14 @@ public class StreamTypeNames {
      */
     public static final String DETECTIONS = "Detections";
 
+    public static final Set<String> VALID_RECEIVE_TYPE_NAMES = new HashSet<>();
+
+    static {
+        VALID_RECEIVE_TYPE_NAMES.add(RAW_EVENTS);
+        VALID_RECEIVE_TYPE_NAMES.add(RAW_REFERENCE);
+        VALID_RECEIVE_TYPE_NAMES.add(EVENTS);
+        VALID_RECEIVE_TYPE_NAMES.add(REFERENCE);
+        VALID_RECEIVE_TYPE_NAMES.add(TEST_EVENTS);
+        VALID_RECEIVE_TYPE_NAMES.add(TEST_REFERENCE);
+    }
 }