+++ stroom-lmdb/src/main/java/stroom/lmdb/LmdbEnvFactory.java
@@ -53,7 +53,6 @@ public class LmdbEnvFactory {
                 lmdbConfig);
     }
 
-
     /**
      * @param dir The path where the environment will be located on the filesystem.
      *            Should be local disk and not shared storage.
@@ -80,6 +79,8 @@ public class LmdbEnvFactory {
                 Paths.get(pathCreator.makeAbsolute(pathCreator.replaceSystemProperties(dir))));
     }
 
+    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
     abstract static class AbstractEnvBuilder {
 
         private static final AtomicBoolean HAS_LIBRARY_BEEN_CONFIGURED = new AtomicBoolean(false);
@@ -92,7 +93,7 @@ public class LmdbEnvFactory {
 
         protected int maxReaders = LmdbConfig.DEFAULT_MAX_READERS;
         protected ByteSize maxStoreSize = LmdbConfig.DEFAULT_MAX_STORE_SIZE;
-        protected int maxDbs = LmdbConfig.DEFAULT_MAX_DATABASES;
+        protected int maxDbs = 1;
         protected boolean isReadAheadEnabled = LmdbConfig.DEFAULT_IS_READ_AHEAD_ENABLED;
         protected boolean isReaderBlockedByWriter = LmdbConfig.DEFAULT_IS_READER_BLOCKED_BY_WRITER;
         protected String subDir = null;
@@ -108,7 +109,6 @@ public class LmdbEnvFactory {
                     getLocalDirAsPath(pathCreator, lmdbConfig));
 
             this.maxReaders = lmdbConfig.getMaxReaders();
-            this.maxDbs = lmdbConfig.getMaxDbs();
             this.maxStoreSize = lmdbConfig.getMaxStoreSize();
             this.isReadAheadEnabled = lmdbConfig.isReadAheadEnabled();
             this.isReaderBlockedByWriter = lmdbConfig.isReaderBlockedByWriter();
@@ -129,6 +129,11 @@ public class LmdbEnvFactory {
                 configureLibrary();
                 HAS_LIBRARY_BEEN_CONFIGURED.set(true);
             }
+        }
+
+        public AbstractEnvBuilder withMaxDbCount(final int maxDbCount) {
+            this.maxDbs = maxDbCount;
+            return this;
         }
 
         public AbstractEnvBuilder addEnvFlag(final EnvFlags envFlag) {
@@ -299,19 +304,19 @@ public class LmdbEnvFactory {
 
         public LmdbEnv build() {
 
-            final Path dir;
+            final Path envDir;
             if (subDir != null && !subDir.isBlank()) {
-                dir = localDir.resolve(subDir);
+                envDir = localDir.resolve(subDir);
 
-                LOGGER.debug(() -> "Ensuring existence of directory " + dir.toAbsolutePath().normalize());
+                LOGGER.debug(() -> "Ensuring existence of directory " + envDir.toAbsolutePath().normalize());
                 try {
-                    Files.createDirectories(dir);
+                    Files.createDirectories(envDir);
                 } catch (IOException e) {
                     throw new RuntimeException(LogUtil.message(
-                            "Error creating directory {}: {}", dir.toAbsolutePath().normalize(), e));
+                            "Error creating directory {}: {}", envDir.toAbsolutePath().normalize(), e));
                 }
             } else {
-                dir = localDir;
+                envDir = localDir;
             }
 
             final Env<ByteBuffer> env;
@@ -332,7 +337,7 @@ public class LmdbEnvFactory {
 
                 LOGGER.info("Creating LMDB environment in dir {}, maxSize: {}, maxDbs {}, maxReaders {}, "
                                 + "isReadAheadEnabled {}, isReaderBlockedByWriter {}, envFlags {}",
-                        dir.toAbsolutePath().normalize(),
+                        envDir.toAbsolutePath().normalize(),
                         maxStoreSize,
                         maxDbs,
                         maxReaders,
@@ -341,15 +346,17 @@ public class LmdbEnvFactory {
                         envFlags);
 
                 final EnvFlags[] envFlagsArr = envFlags.toArray(new EnvFlags[0]);
-                env = builder.open(dir.toFile(), envFlagsArr);
+                env = builder.open(envDir.toFile(), envFlagsArr);
             } catch (Exception e) {
                 throw new RuntimeException(LogUtil.message(
                         "Error creating LMDB env at {}: {}",
-                        dir.toAbsolutePath().normalize(), e.getMessage()), e);
+                        envDir.toAbsolutePath().normalize(), e.getMessage()), e);
             }
-            return new LmdbEnv(localDir, env, isReaderBlockedByWriter);
+            return new LmdbEnv(envDir, env, isReaderBlockedByWriter);
         }
     }
+
+    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     /**
      * A builder for when you don't have an {@link LmdbConfig}
@@ -363,6 +370,7 @@ public class LmdbEnvFactory {
             super(pathCreator, tempDirProvider, lmdbLibraryConfig, dir);
         }
 
+        @Override
         public CustomEnvBuilder withMaxDbCount(final int maxDbCount) {
             this.maxDbs = maxDbCount;
             return this;
@@ -414,6 +422,8 @@ public class LmdbEnvFactory {
         }
     }
 
+    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
     /**
      * A builder for when you have an {@link LmdbConfig}
      */
@@ -424,6 +434,12 @@ public class LmdbEnvFactory {
                                  final LmdbLibraryConfig lmdbLibraryConfig,
                                  final LmdbConfig lmdbConfig) {
             super(pathCreator, tempDirProvider, lmdbLibraryConfig, lmdbConfig);
+        }
+
+        @Override
+        public SimpleEnvBuilder withMaxDbCount(final int maxDbCount) {
+            this.maxDbs = maxDbCount;
+            return this;
         }
 
         @Override