+++ stroom-processor/stroom-processor-impl-db-jooq/src/main/java/stroom/processor/impl/db/jooq/tables/ProcessorFilter.java
@@ -15,7 +15,7 @@ import org.jooq.Identity;
 import org.jooq.Index;
 import org.jooq.Name;
 import org.jooq.Record;
-import org.jooq.Row14;
+import org.jooq.Row16;
 import org.jooq.Schema;
 import org.jooq.Table;
 import org.jooq.TableField;
@@ -42,7 +42,7 @@ import stroom.processor.impl.db.jooq.tables.records.ProcessorFilterRecord;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class ProcessorFilter extends TableImpl<ProcessorFilterRecord> {
 
-    private static final long serialVersionUID = -753718749;
+    private static final long serialVersionUID = -461430865;
 
     /**
      * The reference instance of <code>stroom.processor_filter</code>
@@ -126,6 +126,16 @@ public class ProcessorFilter extends TableImpl<ProcessorFilterRecord> {
      * The column <code>stroom.processor_filter.deleted</code>.
      */
     public final TableField<ProcessorFilterRecord, Boolean> DELETED = createField(DSL.name("deleted"), org.jooq.impl.SQLDataType.BOOLEAN.nullable(false).defaultValue(org.jooq.impl.DSL.inline("0", org.jooq.impl.SQLDataType.BOOLEAN)), this, "");
+
+    /**
+     * The column <code>stroom.processor_filter.min_meta_create_time_ms</code>.
+     */
+    public final TableField<ProcessorFilterRecord, Long> MIN_META_CREATE_TIME_MS = createField(DSL.name("min_meta_create_time_ms"), org.jooq.impl.SQLDataType.BIGINT, this, "");
+
+    /**
+     * The column <code>stroom.processor_filter.max_meta_create_time_ms</code>.
+     */
+    public final TableField<ProcessorFilterRecord, Long> MAX_META_CREATE_TIME_MS = createField(DSL.name("max_meta_create_time_ms"), org.jooq.impl.SQLDataType.BIGINT, this, "");
 
     /**
      * Create a <code>stroom.processor_filter</code> table reference
@@ -230,11 +240,11 @@ public class ProcessorFilter extends TableImpl<ProcessorFilterRecord> {
     }
 
     // -------------------------------------------------------------------------
-    // Row14 type methods
+    // Row16 type methods
     // -------------------------------------------------------------------------
 
     @Override
-    public Row14<Integer, Integer, Long, String, Long, String, String, Integer, Integer, String, Integer, Boolean, Boolean, Boolean> fieldsRow() {
-        return (Row14) super.fieldsRow();
+    public Row16<Integer, Integer, Long, String, Long, String, String, Integer, Integer, String, Integer, Boolean, Boolean, Boolean, Long, Long> fieldsRow() {
+        return (Row16) super.fieldsRow();
     }
 }