+++ stroom-processor/stroom-processor-impl/src/main/java/stroom/processor/impl/MockProcessorFilterDao.java
@@ -18,13 +18,6 @@ public class MockProcessorFilterDao implements ProcessorFilterDao, Clearable {
 
     @Override
     public ProcessorFilter create(final ProcessorFilter processorFilter) {
-        return create(processorFilter, null, null);
-    }
-
-    @Override
-    public ProcessorFilter create(final ProcessorFilter processorFilter,
-                                  final Long minMetaCreateMs,
-                                  final Long maxMetaCreateMs) {
         if (processorFilter.getProcessorFilterTracker() == null) {
             processorFilter.setProcessorFilterTracker(new ProcessorFilterTracker());
         }