+++ CHANGELOG.md
@@ -12,6 +12,48 @@ DO NOT ADD CHANGES HERE - ADD THEM USING log_change.sh
 ~~~
 
 
+## [v7.0-beta.156] - 2021-11-05
+
+* Issue **#2542** : Improve autologged searches "raw" JSON.
+
+* Issue **#2534** : Implement ProcessorResource.fetch to fix event log.
+
+* Issue **#2533** : Fix corrupt event format for logout due to NPE.
+
+* Issue **#2530** : Fix issue of explorer tree not refreshing on copy, move and delete when auth is disabled.
+
+* Add wait time debug logging around LMDB locks.
+
+* Increase maxReaders default to 150 for reference data.
+
+* Issue **#2548** : Clear contents of `stroom.search.resultStore.lmdb.localDir` on boot.
+
+* Issue **#2549** : Remove `maxDbs` from `lmdb` config. Remove `readerBlockedByWriter` from `resultStore.lmdb` config.
+
+* Issue **#2544** : Add additional constraints to processor filter instead of using tracker state
+
+
+## [v7.0-beta.155] - 2021-11-02
+
+* No changes. Previous build failed due to netowrking issues.
+
+
+## [v7.0-beta.154] - 2021-11-02
+
+* Issue **#2540** : No longer error when interrupting shard flush process
+
+* Issue **#2541** : Stop finished queries from deleting LMDB envs for all queries.
+
+
+## [v7.0-beta.153] - 2021-11-01
+
+* Issue **#2535** : Added validation to data type names.
+
+* Issue **#2530** : Fixed explorer refresh on copy and move
+
+* Revert accidental whitespace change to V07_00_00_007__meta_retention_tracker.sql
+
+
 ## [v7.0-beta.152] - 2021-10-29
 
 * Issue **#2519** : Added validation for processor filter expressions and more info about filters to the UI.
@@ -3894,7 +3936,11 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.152...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.156...HEAD
+[v7.0-beta.156]: https://github.com/gchq/stroom/compare/v7.0-beta.155...v7.0-beta.156
+[v7.0-beta.155]: https://github.com/gchq/stroom/compare/v7.0-beta.154...v7.0-beta.155
+[v7.0-beta.154]: https://github.com/gchq/stroom/compare/v7.0-beta.153...v7.0-beta.154
+[v7.0-beta.153]: https://github.com/gchq/stroom/compare/v7.0-beta.152...v7.0-beta.153
 [v7.0-beta.152]: https://github.com/gchq/stroom/compare/v7.0-beta.151...v7.0-beta.152
 [v7.0-beta.151]: https://github.com/gchq/stroom/compare/v7.0-beta.150...v7.0-beta.151
 [v7.0-beta.150]: https://github.com/gchq/stroom/compare/v7.0-beta.149...v7.0-beta.150